Conflicts:
	assembler/test/debruijn/edge_graph_test.hpp

+++ assembler/src/debruijn/repeat_resolver.hpp
@@ -115,14 +115,14 @@ void RepeatResolver<Graph>::ResolveVertex( Graph &g, PairedInfoIndex &ind, Verte
 	}
 	INFO("clustering...");
 
-	int right_edge_count = right_set.size();
+	size_t right_edge_count = right_set.size();
 	vector<vector<EdgeId> > edge_list(right_edge_count);
 	DEBUG("Total: " << right_edge_count << "edges");
 	LOG_ASSERT(right_edge_count == right_vector.size(), "Size mismatch");
 	vector<int> colors(right_edge_count);
-	for(int i = 0; i < right_edge_count; i++)
+	for(size_t i = 0; i < right_edge_count; i++)
 		colors[i] = 0;
-	for(int i = 0; i < right_edge_count; i++) {
+	for(size_t i = 0; i < right_edge_count; i++) {
 //TODO Add option to "jump" - use not only direct neighbours(parameter leap in constructor)
 		DEBUG("Seq in edge:" <<g.EdgeNucls(right_vector[i]).str());
 		vector<EdgeId> neighbours = g.NeighbouringEdges(right_vector[i]);
